refactor: WinningLottoMatcher 클래스 제거 및 로직 수정
