+++ src/main/kotlin/lotto/Main.kt
@@ -6,7 +6,6 @@ import lotto.view.ResultView
 fun main() {
     val lottoCreator = LottoCreator()
     val lottoShop = LottoShop(lottoCreator)
-    val winningLottoMatcher = WinningLottoMatcher()
 
     // 주문 생성
     val amount = InputView.getAmount()
@@ -16,9 +15,9 @@ fun main() {
     // 당첨번호 및 보너스번호 생성
     val winNumberInput = InputView.getWinNumberInput()
     val bonusNumber = InputView.getBonusNumber()
-    val winNumbers = lottoCreator.createWinningLotto(winNumberInput, bonusNumber)
+    val winningLotto = lottoCreator.createWinningLotto(winNumberInput, bonusNumber)
 
     // 결과 출력
-    val result = winningLottoMatcher.checkAndGetResult(order, winNumbers)
+    val result = winningLotto.extractWinningResult(order)
     ResultView.printResult(result)
 }
