+++ src/main/kotlin/lotto/domain/WinningResult.kt
@@ -2,7 +2,7 @@ package lotto.domain
 
 import kotlin.math.roundToInt
 
-class WinningResult(
+class WinningResult private constructor(
     val winningMatchCounts: List<LottoStatistics>,
     amount: Int,
 ) {
@@ -16,5 +16,19 @@ class WinningResult(
 
     private fun calculateRevenue(matchCounts: List<LottoStatistics>): Int {
         return matchCounts.sumOf { it.getTotalPrizeMoney() }
+    }
+
+    companion object {
+        fun from(
+            lottoRanks: Map<Rank, Int>,
+            amount: Int,
+        ): WinningResult {
+            val ranks =
+                Rank.entries
+                    .filter { it !== Rank.MISS }
+                    .map { LottoStatistics(lottoRanks[it] ?: 0, it) }
+                    .sortedBy { it.rank.prizeAmount }
+            return WinningResult(ranks, amount)
+        }
     }
 }
